Writing some more JS unit tests! WIP
